Add renderWidget function to utils
