+++ src/utils/ug-utils.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { lazy, Suspense } from "react";
 import { Parser, ProcessNodeDefinitions } from "html-to-react";
 import { graphql, useStaticQuery, Script } from "gatsby";
 
@@ -132,6 +132,23 @@ const ParseText = ({ textContent }) => {
   return parser.parseWithInstructions(textContent, () => true, instructions);
 };
 
+function renderWidget(componentName, shouldLazyLoad = false, fallback = null, widget, region) {
+  let WidgetModule;
+
+  if(shouldLazyLoad === true) {
+    const Fallback = fallback ? lazy(() => import(`components/shared/${fallback}`)) : () => <></>;
+    WidgetModule = lazy(() => import(`components/shared/${componentName}`));
+    return (
+      <Suspense key={`suspend-${widget.drupal_id}`} fallback={<Fallback />}>
+        <WidgetModule key={widget.drupal_id} data={widget} region={region} />
+      </Suspense>
+    );
+  }
+
+  WidgetModule = require(`components/shared/${componentName}`).default;
+  return <WidgetModule key={widget.drupal_id} data={widget} region={region} />
+}
+
 function setHeadingLevel(headingLevel) {
   // console.log(headingLevel, "setHeadingLevel")
   const validHeadingLevels = ["h1", "h2", "h3", "h4", "h5", "h6"];
@@ -172,6 +189,7 @@ export {
   getNextHeadingLevel,
   isExternalURL,
   ParseText,
+  renderWidget,
   setHeadingLevel,
   slugify,
   stripHTMLTags,
